Move stores out of ActivityLogItem
Reduces the number of event listeners we have on the app page.
